requirements.txt
streamlit
python-dotenv
supabase-py